% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_summary__main.R
\name{ee_summary}
\alias{ee_summary}
\title{Run a pre-specified summary scheme}
\usage{
ee_summary(d, scheme = "Hibbing 2023", ...)
}
\arguments{
\item{d}{input data (presumably a data frame)}

\item{scheme}{character. Name of the routine to be executed. Currently the
only option is \code{"Hibbing 2023"}.}

\item{...}{Arguments passed to sub-routine functions}
}
\value{
A data frame whose contents are prepared according to the
  indicated scheme
}
\description{
Pre-specified routines are designed to facilitate replication of methods from
prior studies
}
\examples{

f <- system.file("extdata/TAS1H30182785_2019-09-17.gt3x", package = "read.gt3x")

## Read acceleration data
  accel <- ee_file(f)

## Generate some bogus count data
  cts <- data.frame(
    Timestamp = accel$Timestamp,
    Axis1 = 0, Axis2 = 0, Axis3 = 0,
    valid_status = "Non-Wear"
  )

## Returns NA with a warning about no complete days in the file
  suppressWarnings(ee_summary(
    merge(accel, cts)
  ))

}
\seealso{
Subroutine function(s): \code{\link{hibbing23-summary}}
}
