% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_features.R
\name{generic_features}
\alias{generic_features}
\title{Calculate generic features for model application}
\usage{
generic_features(
  d,
  time_var = "Timestamp",
  x_var = "Accelerometer_X",
  y_var = "Accelerometer_Y",
  z_var = "Accelerometer_Z",
  win_width_sec = 1,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{d}{data frame of data to use for generating predictions}

\item{time_var}{character. Name of the column containing
POSIX-formatted timestamps}

\item{x_var}{character. Name of the X-axis variable}

\item{y_var}{character. Name of the Y-axis variable}

\item{z_var}{character. Name of the Z-axis variable}

\item{win_width_sec}{desired window width for features}

\item{verbose}{logical. Print updates to console?}

\item{...}{currently unused}
}
\value{
A data frame of features in the specified epoch length
}
\description{
Calculate generic features for model application
}
\examples{
if (isTRUE(requireNamespace("read.gt3x"))) {

  f <- system.file("extdata/TAS1H30182785_2019-09-17.gt3x", package = "read.gt3x")
  d <- stats::setNames(
    read.gt3x::read.gt3x(f, asDataFrame = TRUE, imputeZeroes = TRUE),
    c("Timestamp", "Accelerometer_X", "Accelerometer_Y", "Accelerometer_Z")
  )[1:30000, ]

  utils::head(generic_features(d))

}
}
