% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_file_hibbing23.R
\name{hibbing23-file}
\alias{hibbing23-file}
\alias{ee_file_hibbing23}
\alias{agd_hibbing23}
\alias{gt3x_hibing23}
\title{Run the Hibbing 2023 file preparation routine}
\usage{
## Wrapper function:
ee_file_hibbing23(filename, ...)


## Sub-routine functions
agd_hibbing23(filename, verbose = FALSE, ...)
}
\arguments{
\item{filename}{character. Path to the file}

\item{...}{Arguments passed to sub-routine functions}

\item{verbose}{logical. Print updates to console?}
}
\value{
A data frame whose contents are prepared according to the Hibbing
  2023 scheme
}
\description{
Run the Hibbing 2023 file preparation routine
}
\note{
This routine requires the \code{PhysicalActivity},
  \code{PhysActBedRest}, and \code{read.gt3x} packages. Also note that the
  original Hibbing paper used a package called \code{AGread} rather than
  \code{read.gt3x}, and there are some differences. \code{AGread} is
  available on GitHub, but not CRAN--hence, the requirement for
  \code{read.gt3x} here.
}
\keyword{internal}
