% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_summary_hibbing23.R
\name{hibbing23-summary}
\alias{hibbing23-summary}
\alias{ee_summary_hibbing23}
\title{Run the Hibbing 2023 summary scheme}
\usage{
ee_summary_hibbing23(d, verbose = FALSE)
}
\arguments{
\item{d}{input data (presumably a data frame)}

\item{verbose}{logical. Print updates to console?}
}
\value{
A data frame whose contents are prepared according to the
  indicated scheme
}
\description{
Run the Hibbing 2023 summary scheme
}
\examples{

f <- system.file("extdata/TAS1H30182785_2019-09-17.gt3x", package = "read.gt3x")

## Read acceleration data
  accel <- ee_file(f)

## Generate some bogus count data
  cts <- data.frame(
    Timestamp = accel$Timestamp,
    Axis1 = 0, Axis2 = 0, Axis3 = 0,
    valid_status = "Non-Wear"
  )

## Returns NA with a warning about no complete days in the file
  suppressWarnings(ee_summary(
    merge(accel, cts)
  ))

}
