% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_file_hibbing23.R
\name{hibbing23}
\alias{hibbing23}
\alias{ee_file_hibbing23}
\alias{agd_hibbing23}
\alias{gt3x_hibing23}
\title{Run the Hibbing 2023 file preparation routine}
\usage{
## Wrapper function:
ee_file_hibbing23(filename, ...)


## Sub-routine functions
agd_hibbing23(filename, ...)
}
\arguments{
\item{filename}{character. Path to the file}

\item{...}{Arguments passed to sub-routine functions}

\item{verbose}{logical. Print updates to console?}
}
\value{
A data frame whose contents are prepared according to the Hibbing
  2023 scheme
}
\description{
Run the Hibbing 2023 file preparation routine
}
\note{
This routine requires the \code{PhysicalActivity},
  \code{PhysActBedRest}, and \code{AGread} packages
}
\keyword{internal}
