% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Montoye.R
\name{montoye_features}
\alias{montoye_features}
\title{Calculate features for Montoye's neural networks}
\usage{
montoye_features(
  d,
  time_var = "Timestamp",
  x_var = "Accelerometer_X",
  y_var = "Accelerometer_Y",
  z_var = "Accelerometer_Z",
  side = c("left", "right"),
  win_width_sec = 30,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{d}{data frame of data to use for generating predictions}

\item{time_var}{character. Name of the column containing
POSIX-formatted timestamps}

\item{x_var}{character. Name of the X-axis variable}

\item{y_var}{character. Name of the Y-axis variable}

\item{z_var}{character. Name of the Z-axis variable}

\item{side}{character vector or scalar indicating which side-specific wrist
model(s) to implement for \code{"Montoye 2017"}. Can be \code{"left"},
\code{"right"}, or \code{c("left", "right")}}

\item{win_width_sec}{desired window width for features}

\item{verbose}{logical. Print updates to console?}

\item{...}{currently unused}
}
\value{
A dataframe of features for entry into the neural networks
}
\description{
Calculate features for Montoye's neural networks
}
\examples{
if (isTRUE(requireNamespace("read.gt3x"))) {

  f <- system.file("extdata/TAS1H30182785_2019-09-17.gt3x", package = "read.gt3x")
  d <- read.gt3x::read.gt3x(f, asDataFrame = TRUE, imputeZeroes = TRUE)

  montoye_features(d[1:12000, ], "time", "X", "Y","Z", "right")

}
}
\references{
Montoye et al. (2017) \doi{10.1080/1091367X.2017.1337638}
}
