% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelEE.R
\name{accelEE-function}
\alias{accelEE-function}
\alias{accelEE}
\alias{hildebrand}
\title{Predict energy expenditure for accelerometry data}
\usage{
## Wrapper function:

  accelEE(
    d,
    method = c(
      "Crouter 2006", "Crouter 2010", "Crouter 2012", "Hibbing 2018",
      "Hildebrand Linear", "Hildebrand Non-Linear", "Montoye 2017",
      "Staudenmayer Linear", "Staudenmayer Random Forest"
    ),
    verbose = FALSE,
    ...
  )

## Internal applicator functions that the wrapper calls based on
## the value of the `method` argument (external functions listed
## under 'See Also'):

  hildebrand(
    d, age = c("youth", "adult"), monitor = c("ActiGraph", "GENEActiv"),
    location = c("hip", "wrist"), enmo_name = "ENMO",
    time_var = "Timestamp", vo2_floor_mlkgmin = 3, vo2_ceil_mlkgmin = 70,
    ...
  )
}
\arguments{
\item{d}{data frame of data to use for generating predictions}

\item{method}{the method(s) to use}

\item{verbose}{logical. Print updates to console?}

\item{...}{arguments passed to specific applicators. See details}

\item{age}{the age group(s) of desired Hildebrand equation(s) to apply}

\item{monitor}{the monitor being worn by the participant}

\item{location}{the placement of the monitor on the body}

\item{enmo_name}{name of the variable containing Euclidian Norm
Minus One (ENMO) values}

\item{time_var}{character. Name of the column containing
POSIX-formatted timestamps}

\item{vo2_floor_mlkgmin}{minimum allowable oxygen consumption value (in
ml/kg/min). Values lower than this (if any) will be rounded up to it}

\item{vo2_ceil_mlkgmin}{maximum allowable oxygen consumption value (in
ml/kg/min). Values larger than this (if any) will be rounded down to it}
}
\value{
A data frame appended with new columns containing energy
    expenditure predictions
}
\description{
Predict energy expenditure for accelerometry data
}
\details{
This is a wrapper and aggregator for applying different energy
  expenditure prediction methods. Depending on the value(s) specified in the
  \code{method} argument, calls are made to applicator functions (one per
  method). Most applicators require values to be passed in for additional
  variables. Thus, the signature for each applicator function is included
  above, in the \code{usage section}.
}
\examples{
if (isTRUE(requireNamespace("AGread"))) {

  f <- system.file("extdata/example.gt3x", package = "AGread")
  AG <- AGread::read_gt3x(f, parser = "external")$RAW
  AG <- suppressMessages(AGread::collapse_gt3x(AG))

  utils::head(
    accelEE(AG, "Hibbing 2018", algorithm = 1, site = "Right Wrist")
  )

  utils::head(
    accelEE(
      AG, "Hildebrand Linear", age = "adult",
      monitor = "ActiGraph", location = "Wrist"
    )
  )

}

}
\seealso{
\code{\link[TwoRegression]{TwoRegression}}
}
