% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0b_package_doc.R
\docType{package}
\name{accelEE-package}
\alias{accelEE-package}
\title{accelEE: Predict Energy Expenditure from Accelerometer Data}
\description{
This is a package to simplify and integrate the application of various energy
expenditure models. It is a simple purpose but a powerful tool. As models
proliferate, they can be added into this package and benefit from a
centralized location and a single, unified framework (consistent syntax etc.)
that allows for seamless and flexible use of different methods.
}
\details{
The core function is \code{accelEE}, and the \code{\link[=accelEE]{documentation page}}
has extensive detail on what options and capabilities are available. Start
there and
\href{https://github.com/paulhibbing/accelEE/issues/new}{post an issue on GitHub}
if you need additional clarity, find a bug, or wish to suggest a new
feature.
}
