% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_utils.R
\name{generic_features}
\alias{generic_features}
\title{Calculate generic features for model application}
\usage{
generic_features(
  d,
  time_var = "Timestamp",
  verbose = FALSE,
  output_window_secs = 1
)
}
\arguments{
\item{d}{data frame of ActiGraph data (other monitors not currently
supported)}

\item{time_var}{character. Name of the column in \code{d} containing
POSIX-formatted timestamp information}

\item{verbose}{logical. Print updates to console?}

\item{output_window_secs}{The desired epoch length (in seconds) for output}
}
\value{
A data frame of features in the specified epoch length
}
\description{
This is a loose wrapper around \code{AGread::collapse_gt3x}
}
\note{
Currently, input is only accepted for raw ActiGraph accelerometer data
  processed using the \code{AGread} package.
}
\examples{
if (isTRUE(requireNamespace("AGread"))) {

  f <- system.file("extdata/example.gt3x", package = "AGread")
  d <- AGread::read_gt3x(f, parser = "external")$RAW

  utils::head(generic_features(d))

}
}
