% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_utils.R
\name{generic_features}
\alias{generic_features}
\title{Calculate generic features for model application}
\usage{
generic_features(AG, time_var = "Timestamp", verbose = FALSE)
}
\arguments{
\item{AG}{data frame of ActiGraph data (other monitors not currently
supported)}

\item{time_var}{character. Name of the column in \code{AG} containing
POSIX-formatted timestamp information}

\item{verbose}{logical. Print updates to console?}
}
\value{
A data frame of features in one-second epochs
}
\description{
This is a loose wrapper around \code{AGread::collapse_gt3x}
}
\note{
Currently, input is only accepted for raw ActiGraph accelerometer data
  processed using the \code{AGread} package.
}
\examples{
if (isTRUE(requireNamespace("AGread"))) {

  f <- system.file("extdata/example.gt3x", package = "AGread")
  AG <- AGread::read_gt3x(f, parser = "external")$RAW

  head(generic_features(AG))

}
}
