% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Montoye.R
\name{montoye_features}
\alias{montoye_features}
\title{Calculate features for Montoye's neural networks}
\usage{
montoye_features(
  AG,
  time_var = "Timestamp",
  x_var = "Accelerometer_X",
  y_var = "Accelerometer_Y",
  z_var = "Accelerometer_Z"
)
}
\arguments{
\item{AG}{data frame of ActiGraph data (raw samples)}

\item{time_var}{character. Name of the variable in \code{AG} containing
POSIX-formatted timestamp information}

\item{x_var}{character. Name of the X-axis variable}

\item{y_var}{character. Name of the Y-axis variable}

\item{z_var}{character. Name of the Z-axis variable}
}
\value{
A dataframe of features for entry into the neural networks
}
\description{
Calculate features for Montoye's neural networks
}
\examples{
if (isTRUE(requireNamespace("AGread"))) {

  f <- system.file("extdata/example.gt3x", package = "AGread")
  AG <- AGread::read_gt3x(f, parser = "external")$RAW

  head(montoye_features(AG))

}
}
\references{
\href{https://www.tandfonline.com/doi/abs/10.1080/1091367X.2017.1337638?journalCode=hmpe20}{Montoye et al. (2017)}
}
