% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Staudenmayer.R
\name{staudenmayer_features}
\alias{staudenmayer_features}
\title{Calculate features for Staudenmayer models}
\usage{
staudenmayer_features(
  d,
  time_var = "Timestamp",
  x_var = "Accelerometer_X",
  y_var = "Accelerometer_Y",
  z_var = "Accelerometer_Z",
  win_width_sec = 15,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{d}{data frame of data to use for generating predictions}

\item{time_var}{character. Name of the column containing
POSIX-formatted timestamps}

\item{x_var}{character. Name of the X-axis variable}

\item{y_var}{character. Name of the Y-axis variable}

\item{z_var}{character. Name of the Z-axis variable}

\item{win_width_sec}{desired window width for features}

\item{verbose}{logical. Print updates to console?}

\item{...}{currently unused}
}
\value{
Data frame containing features in the specified format
}
\description{
Calculate features for Staudenmayer models
}
\examples{
if (isTRUE(requireNamespace("read.gt3x"))) {

  f <- system.file("extdata/TAS1H30182785_2019-09-17.gt3x", package = "read.gt3x")
  d <- read.gt3x::read.gt3x(f, asDataFrame = TRUE, imputeZeroes = TRUE)

  staudenmayer_features(d[1:12000, ], "time", "X", "Y", "Z")

}
}
\references{
Staudenmayer et al. (2015), \doi{10.1152/japplphysiol.00026.2015}
}
