% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Staudenmayer.R
\name{staudenmayer_features}
\alias{staudenmayer_features}
\title{Calculate features for Staudanmayer models}
\usage{
staudenmayer_features(
  AG,
  samp_freq = 80,
  win_width_sec = 15,
  time_var = "Timestamp",
  x_var = "Accelerometer_X",
  y_var = "Accelerometer_Y",
  z_var = "Accelerometer_Z"
)
}
\arguments{
\item{AG}{data frame of ActiGraph data (raw samples)}

\item{samp_freq}{sampling frequency}

\item{win_width_sec}{desired window width for features (default is 15 sec)}

\item{time_var}{character. Name of the variable in \code{AG} containing
POSIX-formatted timestamp information}

\item{x_var}{character. Name of the X-axis variable}

\item{y_var}{character. Name of the Y-axis variable}

\item{z_var}{character. Name of the Z-axis variable}
}
\value{
Data frame containing features in the specified format
}
\description{
Calculate features for Staudanmayer models
}
\examples{
if (isTRUE(requireNamespace("AGread"))) {

  f <- system.file("extdata/example.gt3x", package = "AGread")
  AG <- AGread::read_gt3x(f, parser = "external")$RAW

  head(staudenmayer_features(AG, 90))

}
}
\references{
\href{https://pubmed.ncbi.nlm.nih.gov/26112238/}{Staudenmayer et al. (2015)}
}
